feat(): cruise canvas v2
