+++ bricks/ai-portal/src/bootstrap.ts
@@ -11,3 +11,4 @@ import "./mcp-tools/index.js";
 import "./chat-stream/index.js";
 import "./ai-employees/index.js";
 import "./ai-agents/index.js";
+import "./data-providers/request-store.js";
