+++ bricks/ai-portal/src/cruise-canvas/legacy-reducers/jobs.ts
